+++ orix/plot/_util.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
+from typing import Tuple, Union
+
 from matplotlib.patches import FancyArrowPatch
 from mpl_toolkits.mplot3d import proj3d
 import numpy as np
@@ -37,3 +39,83 @@ class Arrow3D(FancyArrowPatch):
         xs, ys, zs = proj3d.proj_transform(xs3d, ys3d, zs3d, self.axes.M)
         self.set_positions((xs[0], ys[0]), (xs[1], ys[1]))
         return np.min(zs)
+
+
+def format_labels(
+    v: Union[list, tuple, np.ndarray],
+    brackets: Tuple[str, str] = ("", ""),
+    use_latex: bool = True,
+) -> np.ndarray:
+    r"""Return formatted vector integer labels.
+
+    This function is a convenient way to get nice labels when plotting
+    vectors in the stereographic projection via
+    :meth:`~orix.vector.Vector3d.scatter` or
+    :meth:`~orix.plot.StereographicPlot.text`.
+
+    Parameters
+    ----------
+    v
+        Vector labels in an array-like with the last dimension having a
+        size of 3 or 4. The labels are rounded to the closets integers
+        before being formatted.
+    brackets
+        Left and right parentheses. These are typically () or [] when
+        labeling reciprocal (hkl) or direct [uvw] lattice vectors,
+        respectively. If not given, no parentheses are added. If
+        ``use_latex=True``, then the brackets {} and <> will be escaped
+        if given.
+    use_latex
+        Whether to use LaTeX when formatting the labels. Default is
+        True.
+
+    Returns
+    -------
+    new_labels
+        Array of string labels.
+
+    Examples
+    --------
+    >>> from orix import plot
+    >>> from orix.vector import Vector3d
+    >>> v = Vector3d([[1, 1, 1], [-2, 0, 1], [4, 0, 0], [-4, 0, 0]])
+    >>> plot.format_labels(v.reshape(2, 2).data)
+    array([['$111$', '$\\bar{2}01$'],
+           ['$400$', '$\\bar{4}00$']], dtype='<U11')
+    >>> plot.format_labels(v.data, ("[", "]"), use_latex=False).tolist()
+    ['[111]', '[-201]', '[400]', '[-400]']
+    >>> plot.format_labels(v.data, ("{", "}")).tolist()
+    ['$\\{111\\}$', '$\\{\\bar{2}01\\}$', '$\\{400\\}$', '$\\{\\bar{4}00\\}$']
+    >>> plot.format_labels(v.data, ("{", "}"), use_latex=False).tolist()
+    ['{111}', '{-201}', '{400}', '{-400}']
+    >>> plot.format_labels(v.data, ("<", ">")).tolist()
+    ['$\\left<111\\right>$',
+     '$\\left<\\bar{2}01\\right>$',
+     '$\\left<400\\right>$',
+     '$\\left<\\bar{4}00\\right>$']
+    """
+    if use_latex:
+        start = end = r"$"
+        if brackets == ("<", ">"):
+            brackets = (r"\left<", r"\right>")
+        elif brackets == ("{", "}"):
+            brackets = (r"\{", r"\}")
+    else:
+        start = end = ""
+
+    v = np.asanyarray(v)
+    shape = v.shape
+    v = v.round().astype(int).reshape(-1, shape[-1])
+
+    new_labels = []
+    for label in v:
+        new_label = start + brackets[0]
+        for i in label:
+            if i < 0 and use_latex:
+                new_label += r"\bar{" + str(abs(i)) + r"}"
+            else:
+                new_label += str(i)
+        new_label += brackets[1] + end
+        new_labels.append(new_label)
+
+    return np.asarray(new_labels).reshape(shape[:-1])